Foi criado o script auxiliar_functions para organizar melhor as funções
